Set everything to use the libmowgli list API
libunicorn will only make use of a very small subset of libmowgli's
functionality, so it may seem like a bad idea to introduce dependencies to
it. However, libmowgli is a top-notch library, and the types of places
where people might be using libunicorn are also the types of places
where libmowgli would be a good choice as well.