hook: Private data argument is supplied at call time
This is opposed to the previous approach in which the private data
argument was supplied at hook time.

Rationale: The previous hook calling convention allowed each hook to
associate itself with a void* pointer that was supplied with each call,
and each call was only allowed to be called with a vector of strings. This
severely limited the capabilities of the hook subsystem.

Although the new format is slightly more verbose, it is better-suited
to the sorts of situations for which libunicorn was intended.